chore: extract util functions
