+++ packages/shared/lib/core/layer-2/utils/index.ts
@@ -3,12 +3,14 @@ export * from './evmAddressToAgentID'
 export * from './getAbiForContractType'
 export * from './getAgentBalanceParameters'
 export * from './getDestinationNetworkFromAddress'
+export * from './getCommonEvmTransactionData'
 export * from './getErc20TokenMetadata'
 export * from './getLayer2MetadataForTransfer'
 export * from './getLayer2NetworkFromAddress'
+export * from './getLayer2AssetAllowance'
 export * from './getSmartContractHexName'
 export * from './parseLayer2Metadata'
 export * from './parseLayer2MetadataForTransfer'
 export * from './prepareErc20EvmTransactionData'
 export * from './prepareEvmTransactionToSign'
-export * from './prepareIrc30TransactionData'
+export * from './signIscpTransferTransactionData'
